그만할래요 살려주세요

+++ src/main/java/kroryi/bus2/service/route/RouteFinder.java
@@ -88,7 +88,7 @@ public class RouteFinder {
                     .transferCount(0)
                     .transferStationId(null)
                     .transferStationName(null)
-                    .estimatedMinutes(linkList.size() * 2)
+                    .estimatedMinutes((double) (linkList.size() * 2))
                     .stationIds(stops)
                     .orsPath(orsPath) // 🔥 여기에 주입
                     .build();
